+++ VCore/Services and Managers/Network/Result/NetworkError.swift
@@ -19,6 +19,9 @@ public enum NetworkError: VCoreError {
     /// An indication that endpoint url is invalid.
     case invalidEndpoint
     
+    /// An indication that request has timed out.
+    case requestTimeOut
+    
     /// An indication that parameters cannot be encoded.
     case incompleteParameters
     
@@ -40,12 +43,13 @@ public enum NetworkError: VCoreError {
     
     public var code: Int {
         switch self {
-        case .notConnectedToNetwork: return 1
-        case .invalidEndpoint: return 2
-        case .incompleteParameters: return 3
-        case .returnedWithError: return 4
-        case .invalidResponse: return 5
-        case .incompleteData: return 6
+        case .notConnectedToNetwork: return 1001
+        case .invalidEndpoint: return 1002
+        case .incompleteParameters: return 1003
+        case .requestTimeOut: return 1004
+        case .returnedWithError: return 1005
+        case .invalidResponse: return 1006
+        case .incompleteData: return 1007
         }
     }
     
@@ -54,6 +58,7 @@ public enum NetworkError: VCoreError {
         case .notConnectedToNetwork: return "Not connected to network"
         case .invalidEndpoint: return "Cannot connect to the server. An incorrect handler is being used."
         case .incompleteParameters: return "Data cannot be encoded or is incomplete"
+        case .requestTimeOut: return "Request has timed out"
         case .returnedWithError: return "Server has encountered an error"
         case .invalidResponse: return "Server has returned an invalid response"
         case .incompleteData: return "Data cannot be decoded or is incomplete"